call_server: check if hash actually exists in v_getProofReceipt
Then, if it does, remove the entry if and only if it is ready
(ProofStatus::Ready(..)). Otherwise, we may end up in a situation where
our task map is lacking the hash that is currently in preflight or in
proving but reporting it as missing instead. I wish there was a way to
acquire the task map's lock for the entire duration of
`v_get_proof_receipt` handler but this seems like an ok alternative
at least for now.