e2e-test.sh: ensure CLI is built by default
Also, rename `BUILD_BINARIES` to `BUILD_SERVICES` and `ensure_binaries_built`
into `ensure_services_built` so that it is clear we only build services
in that function.

The reason I believe we should have an equivalent for building vlayer CLI
only is that it is required in several places to set up a test example
(`vlayer init`) but it may the case that services are run using Docker
in which case if we bundled CLI building into `ensure_binaries_built` we
would unnecessarily build all binaries whereas only CLI is then required.