call_server: move v_allocateGas into task queue
Currently, when the prover gets hit twice with the same proof request,
we will call gas-meter twice which will lead to the gas-meter trying
to write the same record twice into the DB causing a server error.

Instead, we should call the gas-meter as an explicit step in the
queued task worker.

This commit pushes first call to gas-meter into the queued proving task,
and introduces another proving state, AllocateGas, which will report an
error if allocating gas failed in the gas-meter.