feat(cli,jwt): optionally validate JWT signature if pubkey specified
Until now, we always required the public key to be specified which is a little
cumbersome and not always desired. Instead, a better UX in my opinion
would be to optionally require the public key and if not present,
dump the JWT contents *without* signature validation. Then, if the key
is present, we would additionally validate the signature also.